+++ pm4py/algo/simulation/playout/petri_net/variants/basic_playout.py
@@ -22,13 +22,14 @@ class Parameters(Enum):
     MAX_TRACE_LENGTH = "maxTraceLength"
     PETRI_SEMANTICS = "petri_semantics"
     ADD_ONLY_IF_FM_IS_REACHED = "add_only_if_fm_is_reached"
+    FM_LEQ_ACCEPTED = "fm_leq_accepted"
 
 
 def apply_playout(net, initial_marking, no_traces=100, max_trace_length=100,
                   case_id_key=xes_constants.DEFAULT_TRACEID_KEY,
                   activity_key=xes_constants.DEFAULT_NAME_KEY, timestamp_key=xes_constants.DEFAULT_TIMESTAMP_KEY,
                   final_marking=None, return_visited_elements=False, semantics=petri_net.semantics.ClassicSemantics(),
-                  add_only_if_fm_is_reached=False):
+                  add_only_if_fm_is_reached=False, fm_leq_accepted=False):
     """
     Do the playout of a Petrinet generating a log
 
@@ -54,6 +55,8 @@ def apply_playout(net, initial_marking, no_traces=100, max_trace_length=100,
         Semantics of the Petri net to be used (default: petri_net.semantics.ClassicSemantics())
     add_only_if_fm_is_reached
         Adds the case only if the final marking is reached
+    fm_leq_accepted
+        Accepts traces ending in a marking that is a superset of the final marking
     """
     # assigns to each event an increased timestamp from 1970
     curr_timestamp = 10000000
@@ -82,7 +85,7 @@ def apply_playout(net, initial_marking, no_traces=100, max_trace_length=100,
             if not semantics.enabled_transitions(net, marking):  # supports nets with possible deadlocks
                 break
             all_enabled_trans = semantics.enabled_transitions(net, marking)
-            if final_marking is not None and marking == final_marking:
+            if final_marking is not None and final_marking <= marking and (final_marking == marking or fm_leq_accepted):
                 trans = choice(list(all_enabled_trans.union({None})))
             else:
                 trans = choice(list(all_enabled_trans))
@@ -97,9 +100,11 @@ def apply_playout(net, initial_marking, no_traces=100, max_trace_length=100,
 
         if not add_only_if_fm_is_reached:
             all_visited_elements.append(tuple(visited_elements))
-        elif marking == final_marking:
+        elif final_marking == marking:
             all_visited_elements.append(tuple(visited_elements))
-        
+        elif final_marking <= marking and fm_leq_accepted:
+            all_visited_elements.append(tuple(visited_elements))
+
         i = i + 1
 
     if return_visited_elements:
@@ -142,6 +147,7 @@ def apply(net: PetriNet, initial_marking: Marking, final_marking: Marking = None
             Parameters.MAX_TRACE_LENGTH -> Maximum trace length
             Parameters.PETRI_SEMANTICS -> Petri net semantics to be used (default: petri_nets.semantics.ClassicSemantics())
             Parameters.ADD_ONLY_IF_FM_IS_REACHED -> adds the case only if the final marking is reached
+            Parameters.FM_LEQ_ACCEPTED -> Accepts traces ending in a marking that is a superset of the final marking
     """
     if parameters is None:
         parameters = {}
@@ -154,8 +160,10 @@ def apply(net: PetriNet, initial_marking: Marking, final_marking: Marking = None
     return_visited_elements = exec_utils.get_param_value(Parameters.RETURN_VISITED_ELEMENTS, parameters, False)
     semantics = exec_utils.get_param_value(Parameters.PETRI_SEMANTICS, parameters, petri_net.semantics.ClassicSemantics())
     add_only_if_fm_is_reached = exec_utils.get_param_value(Parameters.ADD_ONLY_IF_FM_IS_REACHED, parameters, False)
+    fm_leq_accepted = exec_utils.get_param_value(Parameters.FM_LEQ_ACCEPTED, parameters, False)
 
     return apply_playout(net, initial_marking, max_trace_length=max_trace_length, no_traces=no_traces,
                          case_id_key=case_id_key, activity_key=activity_key, timestamp_key=timestamp_key,
                          final_marking=final_marking, return_visited_elements=return_visited_elements,
-                         semantics=semantics, add_only_if_fm_is_reached=add_only_if_fm_is_reached)
+                         semantics=semantics, add_only_if_fm_is_reached=add_only_if_fm_is_reached,
+                         fm_leq_accepted=fm_leq_accepted)